% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start_CardioRVAR_Shiny_app.R
\name{StartCardioRVARapp}
\alias{StartCardioRVARapp}
\title{Start CardioRVARapp}
\usage{
StartCardioRVARapp()
}
\description{
Launches CardioRVARapp in a browser
}
\examples{
\dontrun{
StartCardioRVARapp()
}


}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
